Register&login&JWT Authentication

+++ backend/controllers/goalController.js
@@ -1,5 +1,6 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 
 // const getGoals = async (req, res) => {
@@ -10,7 +11,7 @@ const Goal = require('../models/goalModel');
 // @route   GET /api/v1/goals
 // @access  Private
 const getGoals = asyncHandler (  async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({user: req.user.id});
     res.json(goals);
 });
 
@@ -23,7 +24,8 @@ const setGoal = asyncHandler( async (req, res) => {
         throw new Error('text is required');
     }
     const goal = await Goal.create({
-        'text' : req.body.text
+        'text' : req.body.text,
+        user: req.user.id
     });
     res.status(200).json({
         goal
@@ -39,7 +41,16 @@ const updateGoal =asyncHandler( async (req, res) => {
         res.status(404)
         throw new Error('Goal not found');
     }
-    console.log(goal);
+    const user = await User.findById(req.user.id);
+    // check for user
+    if(!user) {
+        res.status(404)
+        throw new Error('User not found');
+    }
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error(' user Not authorized');
+    }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
         runValidators: true
@@ -57,6 +68,16 @@ const deleteGoal = asyncHandler( async (req, res) => {
     if(!goal) {
         res.status(404)
         throw new Error('Goal not found');
+    }
+    const user = await User.findById(req.user.id);
+    // check for user
+    if(!user) {
+        res.status(404)
+        throw new Error('User not found');
+    }
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error(' user Not authorized');
     }
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
     res.status(200).json({
